move for_db_tests mod
